// src/types.ts
export type symbolPriceArray = { symbol: string; fullPrice: number }[];
export type coinPaprikaDataArray = {
  symbol: string;
  fullPrice: number;
  price30m: number;
  price1h: number;
  price6h: number;
  price12h: number;
  price24h: number;
}[];
export interface elementType {
  id: string;
  name: string;
  symbol: string;
  rank: number;
  circulating_supply: number;
  total_supply: number;
  max_supply: number;
  beta_value: number;
  first_data_at: string;
  last_updated: string;
  quotes: {
    USD: {
      price: number;
      volume_24h: number;
      volume_24h_change_24h: number;
      market_cap: number;
      market_cap_change_24h: number;
      percent_change_15m: number;
      percent_change_30m: number;
      percent_change_1h: number;
      percent_change_6h: number;
      percent_change_12h: number;
      percent_change_24h: number;
      percent_change_7d: number;
      percent_change_30d: number;
      percent_change_1y: number;
      ath_price: number;
      ath_date: string;
      percent_from_price_ath: number;
    };
  };
}
export interface RequestBody {
  coin: string;
}
export interface RequestApiName {
  api: string;
}
export type CoinType = { id: number; name: string; price: number };